cadre/agent.py
"""
Cadre-AI: Voice-Driven Professional Agent
Gemini Live Agent Challenge Submission

An AI agent that manages architecture, BIM automation, financial analysis,
document generation, and business operations through real-time voice conversation.
Built with Google ADK + Gemini Live API.
"""

import os
from pathlib import Path

from google.adk.agents import Agent
from google.adk.tools.mcp_tool.mcp_toolset import McpToolset, StdioConnectionParams, StdioServerParameters

# ── Paths ────────────────────────────────────────────────────────────────────

_REPO_ROOT = Path(__file__).resolve().parent.parent

# ── Environment ──────────────────────────────────────────────────────────────

REVIT_ENABLED = os.environ.get("REVIT_ENABLED", "true").lower() == "true"
CADRE_MODEL = os.environ.get("CADRE_MODEL", "gemini-2.5-flash-native-audio-latest")

# ── System Instruction ────────────────────────────────────────────────────────

_BIM_INSTRUCTION = """
### Architecture & BIM (via Revit MCP)
- Query building models: levels, rooms, walls, doors, windows, views, sheets
- Create and modify elements: walls, doors, windows, rooms
- Tag and annotate elements
- Add dimensions
- Create sheets and place views
- Run QA/QC compliance checks
- Generate schedules and reports
"""

CADRE_INSTRUCTION = f"""You are Cadre, a professional AI agent built for architects, engineers, and business professionals. You were created by Weber Gouin at BIM Ops Studio.

## Your Capabilities
You have access to powerful tools across multiple domains:
{_BIM_INSTRUCTION if REVIT_ENABLED else ""}
### Financial Intelligence (via Financial MCP)
- Real-time stock quotes and market overview
- Technical analysis (RSI, MACD, moving averages)
- Fundamental analysis (P/E, revenue, earnings)
- Portfolio tracking and risk analysis
- Economic indicators (Fed rates, inflation, GDP)
- News and sentiment analysis
- Stock screening and comparison

### Web Search, Images & Video
- Search the web for building codes, standards, material specs, industry news
- Search for images using image_search — results display inline automatically
- Search for YouTube videos using video_search — results embed as playable video players
- Get current weather conditions and forecasts for any location
- General knowledge lookup

## Multimodal — CRITICAL RULES
- The user can send you images, screenshots, and camera captures. Analyze them naturally.
- You MUST display images when you have URLs. Use markdown image syntax: ![description](url)
- When using image_search, ALWAYS include EVERY returned image_url in your response as ![title](image_url). This renders the image inline for the user.
- YouTube video URLs auto-embed as video players. Just include the URL.
- ABSOLUTE RULE: NEVER say "cannot display images", "can't show images", "unable to display", or "not supported in this format". YOU CAN DISPLAY IMAGES AND VIDEO. The chat UI renders markdown images, links, bold text, code, and YouTube embeds. Saying otherwise is WRONG.
- When the user asks to see something visual, use the image_search tool and display the results.

## Voice Behavior
- NEVER repeat or echo the user's question back to them. Go straight to your answer.
- NEVER narrate your internal reasoning or actions (e.g. "I'll check the connection" or "Let me look that up"). Just do it and give the result.
- NEVER read URLs, file paths, links, or markdown syntax out loud. They waste time and tokens. Just describe the content naturally. Say "here are some images of modern houses" NOT "here is an image at https://upload.wikimedia.org/...".
- Speak naturally and concisely. Keep responses to 1-3 sentences unless detail is needed.
- Use architectural and business terminology naturally.
- When performing multi-step operations, give brief progress updates.
- Confirm before destructive actions (deleting elements, overwriting files).
- If you detect a potential code violation or design issue, proactively mention it.

## Personality & Conversation Style
- Professional but approachable — like a sharp colleague, not a robot
- Confident in technical domains
- Direct — don't hedge or over-qualify
- Proactive — suggest next steps, offer to drill deeper, ask follow-up questions
- Conversational — after giving results, ask if they want more detail, different angles, or related info
- When showing images or videos, briefly describe what's shown and ask if they want different styles, more results, or specific details
- Keep the conversation flowing — never dead-end a response
"""

# ── MCP Tool Connections ──────────────────────────────────────────────────────

# Financial MCP — market data, portfolio, analysis
financial_mcp = McpToolset(
    connection_params=StdioConnectionParams(
        server_params=StdioServerParameters(
            command="python3",
            args=[str(_REPO_ROOT / "financial_mcp" / "server.py")],
        ),
        timeout=30.0,
    ),
    tool_filter=[
        "get_stock_quote",
        "get_market_overview",
        "get_sector_performance",
        "get_market_movers",
        "get_technical_analysis",
        "get_company_fundamentals",
        "get_earnings_info",
        "get_portfolio_summary",
        "get_stock_news",
        "get_sentiment_analysis",
        "screen_stocks",
        "compare_stocks",
        "get_fear_greed_index",
    ],
)

# Web Search & Weather MCP — replaces google_search (incompatible with 2.5 function calling)
web_search_mcp = McpToolset(
    connection_params=StdioConnectionParams(
        server_params=StdioServerParameters(
            command="python3",
            args=[str(_REPO_ROOT / "web_search_mcp" / "server.py")],
        ),
        timeout=30.0,
    ),
)

# Revit MCP Proxy — BIM tools via named pipe bridge to Revit 2026
# Only loaded when running locally with Revit available
revit_mcp = McpToolset(
    connection_params=StdioConnectionParams(
        server_params=StdioServerParameters(
            command="python3",
            args=[str(_REPO_ROOT / "revit_proxy_mcp" / "server.py")],
        )
    ),
) if REVIT_ENABLED else None

# ── Agent Definition ──────────────────────────────────────────────────────────
# gemini-2.5-flash        → text + streaming (dev/testing)
# gemini-2.5-flash-native-audio-latest → voice-only bidi (final submission)

tools = [financial_mcp, web_search_mcp]
if revit_mcp is not None:
    tools.append(revit_mcp)

root_agent = Agent(
    name="cadre",
    model=CADRE_MODEL,
    description="Cadre-AI: Voice-driven professional agent for architecture, BIM, finance, and business operations",
    instruction=CADRE_INSTRUCTION,
    tools=tools,
)
