requirements.txt
google-adk>=1.25.0
google-genai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
mcp>=1.26.0
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.32.0
edge-tts>=7.0.0
